# expression: implement vectorized evaluation for `builtinStrToDateDateSig`