PCP #12101 

### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement vectorized evaluation for `builtinStrToDateDateSig` at #12101 

### What is changed and how it works?
It gets faster aboult 5%.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
```
> GO111MODULE=on go test -check.f TestVectorizedBuiltinTimeFunc
PASS: builtin_time_vec_test.go:246: testEvaluatorSuite.TestVectorizedBuiltinTimeFunc	0.165s
PASS: builtin_time_vec_generated_test.go:192: testEvaluatorSuite.TestVectorizedBuiltinTimeFuncGenerated	0.102s
OK: 2 passed
PASS
ok  	github.com/pingcap/tidb/expression	0.290s
```

 - Benchmark
```
> go test -v -benchmem -bench=BenchmarkVectorizedBuiltinTimeFunc -run=BenchmarkVectorizedBuiltinTimeFunc -args builtinStrToDateDateSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinTimeFuncGenerated-12    	1000000000	         0.0165 ns/op	       0 B/op	       0 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDateSig-VecBuiltinFunc-12         	    2520	    468740 ns/op	   48279 B/op	    1464 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDateSig-NonVecBuiltinFunc-12      	    2402	    500521 ns/op	   48194 B/op	    1464 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDateSig-VecBuiltinFunc#01-12      	     405	   2743202 ns/op	  685807 B/op	   23552 allocs/op
BenchmarkVectorizedBuiltinTimeFunc/builtinStrToDateDateSig-NonVecBuiltinFunc#01-12   	     331	   3618155 ns/op	  685857 B/op	   23552 allocs/op
PASS
ok  	github.com/pingcap/tidb/expression	5.622s
```